Refactored SemanticRelationsStats
List of changes
1. Instead of using semantic chains we started to use semantic lists to evaluate semantic relations
2. Refactored SemanticRelationsStats
3. Method get_semantic_relations_stats is not completed, we should change it to make it valid
4. Added new class — Semantic List, which is used in class SemanticRelationsStats

Description of new way to evaluate semantic relations:
We tried to use different way to evaluate semantic relations, because there is no clear code in previous method. Subsequently, a new method was needed. There are two approaches: by semantic lists and by semantic spaces.

The first way is easier, it is about creating a list of semantically related word for each word in the text. Then, by refining these lists to valuable lists we may calculate the average relationship between amount of unique semants and amount of their usage in a text.

The second is harder, it is about extracting from a text a list of words that semantically connected with each other. It is more looks like semantic space.

Description of validity problem:
After completing refining semantic lists by the code it was expected to give different number of the average relationship between low- and high-scored essays, but it did not. It may be caused by semantic lists with the low relationship because these lists are usually short (2~ words in the list except hypersemant or word that construct the list).